Initial import of luasocket.
This release is work in progress. It has been tested on WinXP, Mac OS X,
SunOS and Linux.

In this version, all Lua code should be built into the binary. For that, you
will need a working versions of luac and bin2c, both available with your
Lua distribution. Check the makefile for details.

==TODO==

- Install files/XXX into correct directory.
- Modify files/XXX to get the correct libpath.
- Question...install the rest of .lua files (http,etc.)

- The correct path must be : ${PREFIX}/share/lua50/