Updated atasm to 1.04.
This version adds a few new features, including the ability to set the
value of the fill byte when generating a raw cartridge image. This
release will also attempt multiple passes to resolve forward
referenced labels that are resized (in Mac/65 this is the dreaded
PHASE ERROR problem).  Version 1.04 fixes a nasty macro bug that
prevented many macros from compiling. Due to the seriousness of this
bug, all users are encouraged to upgade as soon as possible.