Initial import of lua-pgsql
DESCR:
LuaSQL is a simple interface from Lua to a DBMS. It enables a Lua program to:
  * Connect to ODBC, Oracle, MySQL and PostgreSQL databases;
  * Execute arbitrary SQL statements;
  * Retrieve results in a row-by-row cursor fashion;

TODO:
This package is broken for -ldl OS... This may be patched before going to pkgsrc.
It's also possible to create Oracle, MySQL and ODBC linkages using the same sources,
but I only have postgresql74 installed.

Anybody working on postgresql74? porting to bl3 and update databases/postgresql
it's important.