PEAK is the "Python Enterprise Application Kit". If you develop "enterprise" applications with Python, or indeed almost any sort of application with Python, PEAK may help you do it faster, easier, on a larger scale, and with fewer defects than ever before. The key is component-based development, on a reliable infrastructure.
PEAK is an application kit, and applications are made from components. PEAK
provides you with a component architecture, component infrastructure, and
various general-purpose components and component frameworks for building
applications. As with J2EE, the idea is to let you stop reinventing
architectural and infrastructure wheels, so you can put more time into your
actual application.

But PEAK is different from J2EE: it's a single, free implementation of
simpler API's based on an easier-to-use language that can nonetheless scale
with better performance than J2EE.