Initial import of rssh
DESCR:
rssh is a restricted shell for use with OpenSSH,
allowing only scp and/or sftp. For example, if
you have a server which you only want to allow
users to copy files off of via scp, without
providing shell access, you can use rssh to do that.

MESSAGE:
===========================================================================
$NetBSD$

You should configure PermitUserEnvironment in your sshd_config to no.
If you do not do this, it will be possible for users to circumvent rssh.

===========================================================================

TODO:

*****************************************************
**** ********* ************ ************* ***********

I do not generate the PLIST, theorically this pkg works on Linux,
Solaris,HPUX,AIX and Irix, but not in BSD. Because of wordexp 
function. take a look on: http://www.pizzashack.org/rssh/platforms.shtml

If you're on a BSD system without the wordexp() function, you could work around this by obtaining a copy of the wordexp() function's code from, say, glibc2, and creating your own wordexp.h header. Compile wordexp.c manually, and link it against the other .o files. 

*********** ********* *********** ********** ********
*****************************************************