Initial import of this interesting game.
TODO:
Do we have ${PREFIX}/var/games in mtree?

DESCR:
tenmado is a vertically scrolling, late 1990s style (that is, a massive
number of enemy shots against a smaller-than-it-looks spaceship) shoot 'em
up game. A very accurate collision detection makes it a game of dexterity.
If something looks like a triangle, it is a triangle, not a rectangle of
similar size. 

However, surviving is only 20% of the game. The main feature of tenmado is
the "color chain bonus". You can get a very big score (about 100 times
bigger than a normal enemy-destruction point) by destroying enemies of the
same color successively. It is easy or difficult depending on how greedy you
are.