Let's start working on xmms-1.2.10
For DESCR, see audio/xmms

As far as I can tell, this initial import compiles & works on FreeBSD,
and compiles on NetBSD-stable (working untested).

I only use xmms to play mp3 files (not audio CDs), so this package can be
missing some functionalities - that's why I'm importing it, so we can work
on it together.