Initial import of ppp-lzs package as made by Iain Hibbert and submitted via NetBSD PR pkg/21229.
This package is not tested, but pkglint and I see problems.
I did not change anything.

The DESCRiption follows:

This is a NetBSD decompressor for PPP compatible with the Stac LZS
scheme as described in rfc1974.  The algorithm is apparently covered
by patents held by Hifn in the USA and Europe though I wrote this
code independently with no help from Hifn or anybody associated with
them, and I dont know the details of the patents.  You might want
to consider this 'example' code only if that makes you feel better.

In order to compile this, you will need to have the source sets
installed.