Initial import of libmatroska into pkgsrc-wip.
Matroska aims to become THE Standard of Multimedia Container Formats. It
was derived from a project called MCF, but differentiates from it
significantly because it is based on EBML (Extensible Binary Meta
Language), a binary derivative of XML.

Matroska is designed with the future in mind. It incorporates features
you would expect from a modern container format, like:

- Streamable over internet (HTTP and RTP)
- Fast seeking in the file
- High error recovery
- Menus (like DVDs have)
- Chapter entries
- Selectable subtitle streams
- Selectable audio streams
- Modularly Extendable