initial import of sablevm
==> I NEED PPL TO TEST & WORK ON THIS PKG. it's the unique jvm that works with GNUclasspath.
==> I think that pkgsrc needs more work on Java. kaffe and dotgnu for example had a lot
==> of problems to run under netbsd because of library linking , etc (liek libffi, etc)

SableVM is a portable bytecode interpreter written in C, and
implementing the Java virtual machine specification, second edition.
Its goals are to be reasonably small, fast, and efficient, as well
as providing a well-designed and robust platform for conducting research.

TODO:

A lot of work :(((

I don't know why using gmake it builds, and using pkgsrc it does not.

Will be nice to bl3ize the libffi that doesn't builds on freebsd.

It needs a sablevm-lib, that it's in the same sources. build with it?

etc...