Initial import.
This package is needed for compatibility reasons to compile wip/prothon.

TODO:
$Id$

Unsure about PLIST.

Unsure if it will work on a "clean" system with wip/prothon (the package for
bison-1.35 is stripped), but I have no resources to test this.

	-- dotz@irc.pl


DESCR:
Bison is the GNU replacement for yacc(1).  Some programs depend on
extensions present in Bison.

[WARNING]
This is old version of GNU Bison, needed to compile some packages, imported
as a need for backward compatibility. If you plan to use GNU Bison, get a
recent version of the package from devel/bison.