initial import of xmlindent.
XML Indent is a XML stream reformatter written in ANSI C.
It is analogous to GNU indent.

Unicode

UTF-8 and UTF-16 encoded streams will not be indented correctly due
to ASCII-only scanner generated by Flex. There's a patch to Flex,
however, for the brave. 


Problem: Take a look on Makefile:



# XXX *BSD have -lfl in the base...solaris,irix,linux ? propbably they need that
#DEPENDS=		flex>2.5:../../devel/flex