Use buildlink3 instead of buildlink2.
Fix INFODIR to use PREFIX to it too. (Forgot to mention
in earlier log message, but the info file was removed from PLIST
because it already is handled.)

(Forgot to mention in earlier log message, but gettext-lib buildlink
was added because this used gettext() functions. This probably
doesn't matter since this is used under Linux and Linux doesn't
needed libintl.)