RT is an industrial-grade ticketing system. It lets a group of people intelligently and efficiently manage requests submitted by a community of users. RT is used by systems administrators, customer support staffs, NOCs, developers and even marketing departments at over a thousand sites around the world.
Note: rt2 is already in pkgsrc (as devel/rt-mysql and devel/rt-pgsql),
this is rt3. When ready, this should probably be imported as rt-pgsql
and the already pkgs should be renamed to rt2-mysql and rt2-pgsql.
They shouldn't be deleted, because there are a lot of differences
between rt2 and rt3 and many people complain about rt3 being much
slower.