Inkscape is an open source Scalable Vector Graphics (SVG) editor with capabilities similar to Illustrator, CorelDraw, and Visio. Supported SVG features include basic shapes, paths, text, alpha blending, transforms, gradients, node editing, SVG-to-PNG export, grouping, and more.
The goal is to be a fully XML, SVG, and CSS2 compliant extensible
drawing tool. It has a GNOME-style interface.

Inkspace can be used for slideshows, printing, converting image
formats, and can be used without a GUI for some features.

This package also includes inkview, a slideshow program which uses
SVG files.

TODO: maybe make this build under NetBSD with gcc 2.95.3.