A tool to quickly ping N number of hosts to determine their reachability without flooding the network.
fping is different from ping in that you can specify any number of
hosts on the command line, or specify a file containing the lists
of hosts to ping. Instead of trying one host until it timeouts or
replies, fping will send out a ping packet and move on to the next
host in a round-robin fashion. If a host replies, it is noted and
removed from the list of hosts to check. If a host does not respond
within a certain time limit and/or retry limit it will be considered
unreachable.

Unlike ping, fping is meant to be used in scripts and its
output is easy to parse.

This version of fping has IPv6, source address selection and interfaces
support. Notes that this version is still more enchanced from the version 
available on fping.com because of this.