Initial import of the native JDK 1.4.2.
This is Sun's Java[tm] Development Kit, version 1.4.2, made buildable
and usable natively on NetBSD/i386 by Greg Lewis and a host of others.

===== From the patchset README:

This is a set of source code patches for the JDK 1.4.2 SCSL source.
The resulting JDK should be suitable for most uses, although using it in
a production environment is at your own risk.  Please note that this
patchset only provides support for recent versions of FreeBSD and NetBSD.
The patchset is available at
http://www.eyesbeyond.com/freebsddom/java/jdk14.html

Please read the file THANKS for attributions.

Instructions to build the JDK are contained in the file BUILD.

You should also examine the ChangeLog for the changes that have been
made so far, and notes on them.