Initial import.
DESCR:
===========================================================================
Skippy is what is best described as a full-screen task-switcher for X11. It
tries to provide an alternative when taskbars or regular task-switchers
aren't the most efficient way of switching tasks (like when you have a lot
of applications open). When activated (currently only through a hotkey), it
will arrange and scale snapshots of all windows on the current desktop and
it'll let you pick a window using a mouse or a keyboard.

It is pretty similar to a tool made by Apple, called Expose -
	http://www.apple.com/macosx/features/expose/
	
Skippy depends on a NetWM compliant window-manager. So far, it's been tested
using OpenBox-3.2, Fluxbox-0.9.9 and XFWM4 -- but BlackBox, ion, and
WindowMaker aren't supported, although the WindowMaker team is working on
NetWM compliance.