Initial import of freedroid 1.0.2.
Freedroid is a clone of the classic game "Paradroid" on Commodore
64 with some improvements and extensions to the classic version.

In this game, you control a robot, depicted by a small white ball
with a few numbers within an interstellar spaceship consisting of
several decks connected by elevators.

The aim of the game is to destroy all enemy robots, depicted by
small black balls with a few numbers, by either shooting them or
seizing control over them by creating connections in a short subgame
of electric circuits.