Import cube_2004_05_22.
Cube is an open source multiplayer and singleplayer first person shooter
game built on an entirely new and very unconventional engine.  Cube is a
landscape-style engine that pretends to be an indoor FPS engine, which
combines very high precision dynamic occlusion culling with a form of
geometric mipmapping on the whole world for dynamic LOD for configurable
fps & graphic detail on most machines.  Uses OpenGL & SDL.