Initial import of dvdauthor-0.6.10: Set of tools to help you author a DVD.
dvdauthor is a simple set of tools to help you author a DVD.  The idea is to
be able to create menus, buttons, chapters, etc, or you can just take a mpeg
stream and write it to a DVD.

--
Addresses request on {current,netbsd}-users by Brad du Plessis.  Untested.