S.C.O.U.R.G.E.: Vermin Extermination Services.  Graphical 3D roguelike game.
``Life is comfortable in the quasi-medieval city of Horghh. In fact it
is so comfortable that a group of aging adventurers, having no monsters
to slay, young people to rescue or tree-stranded pets to fetch, decide
to incorporate into a dungeon janitorial services company. After
renting a run-down office-hovel on the shady side of town, the group
accepts missions from a line of increasingly dangerous low-lifes. Our
heros are hired by overlords, evil sorcerers and underworld bosses to
exterminate pests occupying old dungeons, haunted towers, and other
future centers of world domination. While working for the daily mutton,
group members battle alcohol, decrese of self- confidence and loss of
integrity. It is an epic battle of mental and psychic proportions.''