Scribus is a desktop page layout program in the tradition of Corel Ventura, Quark Xpress, PageMaker and InDesign.
Since its launch in the spring of 2001, Scribus 1.+ offers Linux and Unix
users a versatile and user friendly page layout application. Scribus 1.0 and
its recent development versions are being used in a number of ways; from
brochure design to newsletters and posters to technical documentation.
Scribus has the type of the features one would expect in a sophisticated
page layout application. You can do all the typical tasks like precision
placing and rotating of text and/or images on a page, specify manual
kerning of type and much more. With the release of Scribus 1.0, Linux and
Unix users now have one more high quality application for the desktop,
making it the premier choice for DTP on Linux or BSD with other
platforms to come.