Initial import of yabause-0.0.6: Yabause is a Sega Saturn emulator under GNU GPL.
Be aware that this project is at an early stage. Don't expect to
play commercial games on your PC. This software is released for
testing purposes only, and is essentially aimed to developers.


Imported into wip since the program will probably take quite
some more time until it is usable.