T-Gnus (gnu with T-Shirt) is like the larsi Gnus built into emacs but, uh, different.
It uses SEMI for its MIME support rather than the MIME stuff built into 
regular Gnus.  It's most important feechur is the printing of an annoyingly 
long User-agent: header so that everyone can tell you are running an 
obscure Gnus variant, and also that you're running it under NetBSD.  Also 
it comes with manuals in Japanese.