initial import of ttysnoop
The package allows you to snoop on login tty's through another tty-device
or pseudo-tty. The snoop-tty becomes a 'clone' of the original tty,
redirecting both input and output from/to it.

TODO:

set perms to /var/spool/ttysnoop

test if it works. I was unable to set up a telnet daemon with ttysnoop inside :/