initial import for java-gtk24
Java-GNOME is a set of Java bindings for GNOME and GTK libraries,
so GNOME applications can be written in Java.  It is implemented
as a JNI layer that delegates the Java calls out to the underlying
C library.

This package provides GTK2 bindings for Java(TM).

TODO:

port to kaffe, etc. not tested yet.